// src/types/geolocations.type.ts
export type GeocodingDTO = Geocoding[];

export interface Geocoding {
  name: string;
  local_names?: LocalNames;
  lat: number;
  lon: number;
  country: string;
  state?: string;
}

interface LocalNames {
  ascii?: string;
  feature_name?: string;
  ab?: string;
  af?: string;
  am?: string;
  an?: string;
  ar?: string;
  av?: string;
  ay?: string;
  az?: string;
  ba?: string;
  be?: string;
  bg?: string;
  bh?: string;
  bi?: string;
  bm?: string;
  bn?: string;
  bo?: string;
  br?: string;
  bs?: string;
  ca?: string;
  ce?: string;
  co?: string;
  cr?: string;
  cs?: string;
  cu?: string;
  cv?: string;
  cy?: string;
  da?: string;
  de?: string;
  ee?: string;
  el?: string;
  en?: string;
  eo?: string;
  es?: string;
  et?: string;
  eu?: string;
  fa?: string;
  ff?: string;
  fi?: string;
  fj?: string;
  fo?: string;
  fr?: string;
  fy?: string;
  ga?: string;
  gd?: string;
  gl?: string;
  gn?: string;
  gu?: string;
  gv?: string;
  ha?: string;
  he?: string;
  hi?: string;
  hr?: string;
  ht?: string;
  hu?: string;
  hy?: string;
  ia?: string;
  id?: string;
  ie?: string;
  ig?: string;
  io?: string;
  is?: string;
  it?: string;
  iu?: string;
  ja?: string;
  jv?: string;
  ka?: string;
  kk?: string;
  kl?: string;
  km?: string;
  kn?: string;
  ko?: string;
  ku?: string;
  kv?: string;
  kw?: string;
  ky?: string;
  lb?: string;
  li?: string;
  ln?: string;
  lo?: string;
  lt?: string;
  lv?: string;
  mg?: string;
  mi?: string;
  mk?: string;
  ml?: string;
  mn?: string;
  mr?: string;
  ms?: string;
  mt?: string;
  my?: string;
  na?: string;
  ne?: string;
  nl?: string;
  nn?: string;
  no?: string;
  nv?: string;
  ny?: string;
  oc?: string;
  oj?: string;
  om?: string;
  or?: string;
  os?: string;
  pa?: string;
  pl?: string;
  ps?: string;
  pt?: string;
  qu?: string;
  rm?: string;
  ro?: string;
  ru?: string;
  sa?: string;
  sc?: string;
  sd?: string;
  se?: string;
  sh?: string;
  si?: string;
  sk?: string;
  sl?: string;
  sm?: string;
  sn?: string;
  so?: string;
  sq?: string;
  sr?: string;
  st?: string;
  su?: string;
  sv?: string;
  sw?: string;
  ta?: string;
  te?: string;
  tg?: string;
  th?: string;
  tk?: string;
  tl?: string;
  to?: string;
  tr?: string;
  tt?: string;
  tw?: string;
  ug?: string;
  uk?: string;
  ur?: string;
  uz?: string;
  vi?: string;
  vo?: string;
  wa?: string;
  wo?: string;
  yi?: string;
  yo?: string;
  zh?: string;
  zu?: string;
}